feat(testing): extract testSaga into new workspace package
