+++ packages/saga/src/index.ts
@@ -80,15 +80,6 @@ export {
   type SagaTriggerDefinition
 } from './createSaga';
 export {
-  testSaga,
-  type TestSagaFixture,
-  type TestSagaOptions,
-  type TestSagaInvokeFailure,
-  type TestSagaInvokeFailureReason,
-  type TestSagaInvokeResult,
-  type TestSagaInvokeSuccess
-} from './testSaga';
-export {
   startPolicy,
   type SagaRestartMode,
   type SagaRestartOptions,
